+++ dart/dynamics/InverseKinematics.cpp
@@ -467,7 +467,7 @@ InverseKinematics::TaskSpaceRegion::TaskSpaceRegion(
 std::unique_ptr<InverseKinematics::ErrorMethod>
 InverseKinematics::TaskSpaceRegion::clone(InverseKinematics* _newIK) const
 {
-  return common::make_unique<TaskSpaceRegion>(
+  return std::make_unique<TaskSpaceRegion>(
       _newIK, getTaskSpaceRegionProperties());
 }
 
@@ -816,7 +816,7 @@ InverseKinematics::JacobianDLS::JacobianDLS(
 std::unique_ptr<InverseKinematics::GradientMethod>
 InverseKinematics::JacobianDLS::clone(InverseKinematics* _newIK) const
 {
-  return common::make_unique<JacobianDLS>(_newIK, getJacobianDLSProperties());
+  return std::make_unique<JacobianDLS>(_newIK, getJacobianDLSProperties());
 }
 
 //==============================================================================
@@ -875,7 +875,7 @@ InverseKinematics::JacobianTranspose::JacobianTranspose(
 std::unique_ptr<InverseKinematics::GradientMethod>
 InverseKinematics::JacobianTranspose::clone(InverseKinematics* _newIK) const
 {
-  return common::make_unique<JacobianTranspose>(
+  return std::make_unique<JacobianTranspose>(
       _newIK, getGradientMethodProperties());
 }