+++ tutorials/tutorial_biped_finished/main.cpp
@@ -220,8 +220,7 @@ public:
   {
     setWorld(world);
 
-    mController
-        = dart::common::make_unique<Controller>(mWorld->getSkeleton("biped"));
+    mController = std::make_unique<Controller>(mWorld->getSkeleton("biped"));
   }
 
   /// Handle keyboard input